feat(user): created modal created added creation functionality
